+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/InvokerUtils.java
@@ -41,10 +41,11 @@ import org.codehaus.plexus.logging.Logger;
 import static java.util.Objects.requireNonNull;
 
 /**
- * Various utilities, mostly to bridge "old" and "new" stuff, like Properties vs Maps, File vs Paths, etc.
+ * Various utilities.
  */
-public final class Utils {
-    private Utils() {}
+public final class InvokerUtils
+{
+    private InvokerUtils() {}
 
     @Nonnull
     public static String stripLeadingAndTrailingQuotes(String str) {
