+++ impl/maven-cli/src/main/java/org/apache/maven/cling/invoker/mvn/CommonsCliMavenOptions.java
@@ -33,7 +33,7 @@ import org.apache.maven.cling.invoker.CommonsCliOptions;
 import org.codehaus.plexus.interpolation.BasicInterpolator;
 import org.codehaus.plexus.interpolation.InterpolationException;
 
-import static org.apache.maven.cling.invoker.Utils.createInterpolator;
+import static org.apache.maven.cling.invoker.InvokerUtils.createInterpolator;
 
 public class CommonsCliMavenOptions extends CommonsCliOptions implements MavenOptions {
     public static CommonsCliMavenOptions parse(String source, String[] args) throws ParseException {
